% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiere.mali-package.R
\docType{package}
\name{filiere.mali-package}
\alias{filiere.mali}
\alias{filiere.mali-package}
\title{filiere.mali: Analyse de Filiere Agricole - EHCVM Mali 2021/2022}
\description{
Outils reproductibles pour l'etude de la filiere mil a partir des donnees EHCVM Mali 2021/2022.
}
\author{
\strong{Maintainer}: Aissatou SAMB \email{aissatousamb2@gmail.com}

Authors:
\itemize{
  \item Aissatou SAMB \email{aissatousamb2@gmail.com}
  \item Mourtalla SALL \email{mouslim0234@gmail.com}
}

}
\keyword{internal}
