% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{create_dir}
\alias{load_dta}
\alias{save_rds}
\alias{load_rds}
\alias{check_file}
\alias{winsorize}
\alias{safe_divide}
\title{Utilitaires de chargement et sauvegarde}
\usage{
create_dir(path)

load_dta(folder, filename)

save_rds(object, folder, filename)

load_rds(folder, filename)

check_file(file_path)

winsorize(x, probs = c(0.01, 0.99))

safe_divide(num, den)
}
\arguments{
\item{path}{Chemin du dossier à créer si nécessaire.}

\item{folder}{Dossier contenant le fichier.}

\item{filename}{Nom du fichier.}

\item{object}{Objet R à sauvegarder.}

\item{file_path}{Chemin complet du fichier à vérifier.}

\item{x}{Vecteur numérique.}

\item{probs}{Vecteur de deux probabilités pour la winsorisation.}

\item{num, den}{Numérateur et dénominateur pour une division sécurisée.}
}
\description{
Fonctions pratiques pour charger et sauvegarder des données au format
Stata (.dta) et RDS.
}
